Refactor API client import paths in multiple components and enhance error handling in apiConfig
